refactor(db): extract residency apply engine to db/queries/residency-apply.ts
Both transition directions apply the same per-table rules: attach ingest (T2,
host side) and detach apply (T4, member side). Move the engine — the rule matrix,
ensure-agent, and the per-kind apply functions — out of host/routed-residency.ts
into a DB-layer, transport-free module so the two sides share one implementation
instead of duplicating it.

applyResidencyRows(db, table, rows, { logger }) keeps its exact signature and
ResidencyApplyResult { applied } shape; routed-residency.ts becomes the thin
HTTP/validation/transaction shell that imports it. Behaviour is unchanged (pure
move); tests re-point their engine imports to the new module.
